Move theme and event comments listprovider
