+++ modules/admin-service/src/main/java/org/opencastproject/adminui/endpoint/ListProvidersEndpoint.java
@@ -40,9 +40,9 @@ import org.opencastproject.list.common.query.JobsListQuery;
 import org.opencastproject.list.common.query.SeriesListQuery;
 import org.opencastproject.list.common.query.ServersListQuery;
 import org.opencastproject.list.common.query.ServicesListQuery;
-import org.opencastproject.list.common.query.ThemesListQuery;
 import org.opencastproject.list.common.query.UsersListQuery;
 import org.opencastproject.security.api.SecurityService;
+import org.opencastproject.themes.list.ThemesListQuery;
 import org.opencastproject.util.doc.rest.RestParameter;
 import org.opencastproject.util.doc.rest.RestQuery;
 import org.opencastproject.util.doc.rest.RestResponse;
