+++ modules/admin-service/src/main/java/org/opencastproject/adminui/endpoint/ThemesEndpoint.java
@@ -41,7 +41,6 @@ import org.opencastproject.elasticsearch.index.ElasticsearchIndex;
 import org.opencastproject.elasticsearch.index.objects.series.Series;
 import org.opencastproject.elasticsearch.index.objects.series.SeriesSearchQuery;
 import org.opencastproject.index.service.util.RestUtils;
-import org.opencastproject.list.common.query.ThemesListQuery;
 import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.security.api.User;
@@ -51,6 +50,7 @@ import org.opencastproject.staticfiles.api.StaticFileService;
 import org.opencastproject.staticfiles.endpoint.StaticFileRestService;
 import org.opencastproject.themes.Theme;
 import org.opencastproject.themes.ThemesServiceDatabase;
+import org.opencastproject.themes.list.ThemesListQuery;
 import org.opencastproject.themes.persistence.ThemesServiceDatabaseException;
 import org.opencastproject.util.DateTimeSupport;
 import org.opencastproject.util.EqualsUtil;
