+++ modules/event-comment/pom.xml
@@ -127,6 +127,7 @@
           <instructions>
             <Export-Package>
               org.opencastproject.event.comment;version=${project.version},
+              org.opencastproject.event.comment.list;version=${project.version},
               org.opencastproject.event.comment.persistence;version=${project.version}
             </Export-Package>
             <Import-Package>
