+++ modules/themes/pom.xml
@@ -59,6 +59,18 @@
       <artifactId>h2</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.opencastproject</groupId>
+      <artifactId>opencast-list-providers-api</artifactId>
+      <version>${project.version}</version>
+      <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.opencastproject</groupId>
+      <artifactId>opencast-list-providers-common</artifactId>
+      <version>${project.version}</version>
+      <scope>compile</scope>
+    </dependency>
   </dependencies>
   <build>
     <plugins>
@@ -83,6 +95,7 @@
           <instructions>
             <Export-Package>
               org.opencastproject.themes;version=${project.version},
+              org.opencastproject.themes.list;version=${project.version},
               org.opencastproject.themes.persistence;version=${project.version}
             </Export-Package>
             <Meta-Persistence>
